requirements.txt
scystream-sdk==1.2.1
psycopg2-binary==2.9.10
